refactor: consolidate HTTP test server implementations
Consolidates two duplicate HTTP test server implementations into a single
shared utility that can be used by both OTLP exporter tests and URLSession
instrumentation tests.

Changes:
- Create unified HttpTestServer in Tests/Shared/TestUtils that combines
  functionality from both previous implementations
- Uses POSIX sockets directly, maintaining independence from external
  dependencies (no NIO, Swifter, etc.)
- Supports OTLP test requirements: request recording, protobuf binary
  payloads, header verification
- Supports URLSession test requirements: path-based responses, success/error
  callbacks, configuration options
- Add SharedTestUtils target to Package.swift for shared test utilities
- Update both test targets to depend on SharedTestUtils
- Remove duplicate implementations from ExportersTests and
  InstrumentationTests
- Update test files to import SharedTestUtils and use consistent API

Benefits:
- Eliminates code duplication between test suites
- Provides consistent HTTP mocking functionality
- Makes it easier to maintain and extend test infrastructure
- All OTLP HTTP exporter tests (16) passing successfully

Note: URLSession tests have 4 pre-existing failures in error handling
that are unrelated to this refactoring.
